Add Auth0 integration
[Problem]
Need authentication capabilities for the WAISN tech server.

[Solution]
Integrate with Auth0 and allow disabling authentication via environment settings. Note that the **Log In** button still
exists even with authentication turned off and will redirect you to Auth0. The button could be removed, but I don't
think we want the test and production UI's to look different.

[Test]
* `python manage.py test`
* set authentication on and logged in and out via Auth0
* set authentication off and authentication was not required

In the future, end-2-end tests could be used to automate this using selenium or cypress.